# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: VN Fashion Shop API
  description: |
    RESTful API cho nen tang thuong mai dien tu thoi trang Viet Nam.
    RESTful API for the Vietnamese fashion e-commerce platform.

    ## Xac thuc / Authentication
    Su dung Bearer JWT token trong header `Authorization`.
    Use Bearer JWT tokens in the `Authorization` header.

    ## Phan trang / Pagination
    Cac endpoint tra ve danh sach su dung query params `page` va `page_size`.
    List endpoints use `page` and `page_size` query parameters.

    ## Ma loi / Error Codes
    - `400` -- Bad Request
    - `401` -- Unauthorized (chua xac thuc)
    - `403` -- Forbidden (khong co quyen)
    - `404` -- Not Found
    - `422` -- Validation Error
    - `500` -- Internal Server Error
  version: "1.0.0"
  contact:
    name: VN Fashion Shop Team
    email: dev@vnfashion.vn
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.vnfashion.vn
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token nhan duoc tu endpoint /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Khong tim thay san pham"
...
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          example: "VNF-20260101-0001"
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        subtotal:
          type: number
          format: double
          example: 1200000
        shipping_fee:
          type: number
          format: double
          example: 30000
        discount_amount:
          type: number
          format: double
          example: 0
        total_amount:
          type: number
          format: double
          example: 1230000
        currency:
          type: string
          example: "VND"
        payment_method:
          type: string
          example: "stripe"
        payment_status:
          type: string
          enum: [pending, paid, failed, refunded]
        shipping_name:
          type: string
          example: "Nguyen Van A"
        shipping_phone:
          type: string
          example: "0901234567"
        shipping_address:
          type: string
          example: "123 Nguyen Hue, Quan 1"
        shipping_city:
          type: string
          example: "TP. Ho Chi Minh"
        shipping_province:
          type: string
          example: "Ho Chi Minh"
        tracking_number:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        paid_at:
          type: string
          format: date-time
          nullable: true
        shipped_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_name:
          type: string
          example: "Ao Dai Hoa Sen"
        variant_sku:
          type: string
          example: "ADHS-DO-M"
        size:
          type: string
          example: "M"
        color:
          type: string
          example: "Do"
        unit_price:
          type: number
          format: double
          example: 1200000
        quantity:
          type: integer
          example: 1
        total_price:
          type: number
          format: double
          example: 1200000
    PaginatedOrders:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer

    # ---- Admin ----
    AdminUpdateOrderRequest:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, processing, shipped, delivered, cancelled]
        tracking_number:
          type: string
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: Danh sach don hang cua toi / My orders
      operationId: listMyOrders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
      responses:
        "200":
          description: Danh sach don hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Chi tiet don hang / Order detail
      operationId: getOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chi tiet don hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Khong co quyen xem don hang nay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay don hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{order_id}/cancel:
    post:
      tags: [Orders]
      summary: Huy don hang / Cancel order
      operationId: cancelOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Huy don thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Don hang khong the huy (da giao hoac da huy)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Khong the huy don hang da giao"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay don hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ======== Admin ========
...
  /admin/orders:
    get:
      tags: [Admin]
      summary: Tat ca don hang / All orders
      operationId: adminListOrders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        - name: search
          in: query
          description: Tim theo ma don hang hoac ten khach
          schema:
            type: string
      responses:
        "200":
          description: Danh sach don hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Khong co quyen admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{order_id}:
    patch:
      tags: [Admin]
      summary: Cap nhat trang thai don hang / Update order status
      operationId: adminUpdateOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateOrderRequest"
      responses:
        "200":
          description: Cap nhat thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Khong co quyen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay don hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...